refactor: Modal 하나로 합치기
