+++ frontend/src/types/myPage/types.ts
@@ -3,9 +3,12 @@ import { Dispatch, SetStateAction } from 'react';
 export interface MyPageModalProps {
   showModal: boolean;
   setShowModal: Dispatch<SetStateAction<boolean>>;
-  type: string;
-  navigation: any;
+  type?: string;
+  navigation?: any;
   memberId: number | null;
+  scheduleId?: number | null;
+  handleCheckPlanDetail?: () => void;
+  setFocus?: Dispatch<SetStateAction<boolean>>;
 }
 
 export interface ContentProps {
